:recycle: Better repartition by adding a .cpp file
